Updated CFLAGS and changed compile rules
I made the rules for compiling into two seperate cases. It doesn't change
anything, but it seems more thorough this way. I also added the compiler flag
that changes warnings into errors. I don't know if that is good practice, but I
find it comforting that my code is not breaking due to known issues.