Added shutdown command.
Clients now have the ability to send a 'shutdown' command, that ends the server.
In truth, this should probably be handled server side, and I will make that
shift at some point. The idea was to see if I could have a client give a command
that has the server respond by sending data to each client. Now that it is
sorted out, it all comes down to the creativity of the commands.